fix(KYC): Use ValidationTextField in KYCConfirmPhoneNumberController.
